refactor(structure): app shell + ui/Avatar; biome format pass
- src/app/components/useDevicePushToggle.ts → hooks/
- src/app/components/ward-picker.tsx → WardPicker.tsx (PascalCase)
- src/app/auth-gate.tsx → AuthGate.tsx (PascalCase)
- src/components/ui/Avatar promoted to folder (has a test):
  Avatar/{Avatar.tsx, index.tsx, __tests__/Avatar.test.tsx}
- Run biome format on the tree to settle line-break drift introduced
  by the path rewrites (6 files, no semantic changes).

End-state verification:
  - typecheck clean
  - lint 200 warnings (pre-existing) / 0 errors
  - format:check clean
  - 45 test files / 346 tests pass
  - vite build clean
