+++ src/features/plan-speakers/utils/persistRoster.ts
@@ -1,4 +1,8 @@
-import { createSpeaker, deleteSpeaker, updateSpeaker } from "@/features/speakers/utils/speakerActions";
+import {
+  createSpeaker,
+  deleteSpeaker,
+  updateSpeaker,
+} from "@/features/speakers/utils/speakerActions";
 import type { NonMeetingSunday } from "@/lib/types";
 import { isDirty, type RosterDraft } from "./rosterDraft";
