+++ src/features/program-templates/nodes/VariableChipView.tsx
@@ -2,7 +2,10 @@ import { useEffect, useRef, useState } from "react";
 import { $getNodeByKey, type NodeKey } from "lexical";
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
 import { useLetterVars, useLiveValues } from "@/features/page-editor/utils/letterRenderContext";
-import { useVariableMeta, useVariableRegistry } from "@/features/page-editor/utils/variableRegistry";
+import {
+  useVariableMeta,
+  useVariableRegistry,
+} from "@/features/page-editor/utils/variableRegistry";
 import { styleStringToReact } from "./utils/chipStyle";
 import { VariableChipNode } from "./VariableChipNode";
 import { VariableChipPicker } from "./VariableChipPicker";
